% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_visualization.R
\name{calculate_cluster_ctmeans}
\alias{calculate_cluster_ctmeans}
\title{Calculate isoform-level summary metrics for a cluster}
\usage{
calculate_cluster_ctmeans(data, isoform_ids, id_table, isoform_col = NULL)
}
\arguments{
\item{data}{A data.frame or tibble including single-cell data, with
isoforms as rows and cells as columns. Isoform IDs can be included as row
names (data.frame) or as an additional column (tibble). Expression should be
previously \strong{scaled} (see \code{\link{scale_range}}).}

\item{isoform_ids}{A character vector including the IDs of the isoforms included
in the cluster.}

\item{id_table}{A data frame including two columns named \code{cell} and \code{cell_type},
providing the correspondence between cell ID and cell type should be
provided. The number of rows should be equal to the total number of
cell columns in \code{data}, and the order of the \code{cell} column should
match column (i.e. cell) order in \code{data}.}

\item{isoform_col}{When a tibble is provided in \code{data}, a character
object indicating the name of the column where isoform IDs are specified.
Otherwise, isoform identifiers will be assumed to be defined as rownames,
and this argument will not need to be provided.}
}
\value{
A tibble containing four columns: isoform IDs, mean expression values
in each of the cell types, cell type names, and standard error of cell type-level
expression.
}
\description{
This function computes cell type mean expression and standard
error for all the isoforms in a cluster.
}
