% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{cluster_isoforms}
\alias{cluster_isoforms}
\title{Initial clustering of isoforms using dynamicTreeCut}
\usage{
cluster_isoforms(
  cor_matrix,
  deepSplit = 3,
  pamStage = FALSE,
  minClusterSize = 20,
  ...
)
}
\arguments{
\item{cor_matrix}{A matrix including co-expression values (e.g.
correlations) between isoforms, where column and row names indicate isoform
IDs.}

\item{deepSplit}{Either logical or integer int he range 0 to 4 (defaults to 3).
Provides control over sensitivity to cluster splitting (higher value creates
more clusters with a smaller number of elements).}

\item{pamStage}{A logical indicating whether a second clustering step similar
to Partition Around Medioids is to be performed. If \code{pamStage = TRUE} is
set, no isoforms will remain unassigned to clusters. By default, we allow
unclustered elements by setting \code{pamStage = FALSE}.}

\item{minClusterSize}{An integer defining minimum cluster size allowed
during dynamic clustering. Defaults to 20.}

\item{...}{Additional parameters supplied to
\code{\link[dynamicTreeCut]{cutreeHybrid}}.}
}
\value{
A list of generated clusters, that is, of character vectors including
the identifiers of isoforms assigned to each of the clusters.

\strong{Note:} if \code{PAM = FALSE}, the first element of the list will
correspond to \emph{"cluster zero"}, i.e. a group containing isoforms that
were not assigned to any of the clusters.
}
\description{
Create groups of isoforms with similar expression across cell
types. Based on co-expression values and dynamic
hierarchical clustering with the \code{dynamicTreeCut} package.
Ultimately, this function provides a custom wrapper for the
\code{\link[dynamicTreeCut]{cutreeHybrid}} function.
}
\details{
First, a previously obtained \strong{co-expression matrix} (for instance,
percentile correlations generated using \code{\link{percentile_cor}})
is transformed into a distance matrix for hierarchical clustering.
As the \code{acorde} pipeline intends to detect positively correlated isoforms,
negative correlation values are automatically discarded by replacing them
with zero values. Distance is computed as \emph{1 - |co-expression|}.

Next, \strong{hierarchical clustering} is performed using the
\code{\link[stats]{hclust}} function. Using the output dendrogram, isoforms
are clustered via the \code{\link[dynamicTreeCut]{cutreeHybrid}} function
in the \code{dynamicTreeCut} package.
}
\references{
\insertRef{Langfelder2008}{acorde}

\insertRef{Venables2002}{acorde}
}
\seealso{
For more details on parameters, refer to
\code{\link[dynamicTreeCut]{cutreeHybrid}}.
}
