% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{consensus_DE_set}
\alias{consensus_DE_set}
\title{Isoform Differential Expression analysis results}
\format{
A tibble containing a single column, named \code{transcript}.
This transcript list is used to filter the \code{\link{tasic}} expression
matrix to retain only consistently DE transcript isoforms.
}
\usage{
consensus_DE_set
}
\description{
This object contains a list of differentially expressed isoforms
obtained after x50 downsamplping runs combined with Differential Expression
testing using edgeR and DESeq2. Given potential variability across
DE method runs and to ensure the reproducibility of the results in Arzalluz-Luque
et al. (see refs.), we provide the original results used in the manuscript.
}
\references{
\insertRef{Arzalluz-Luque2021}{acorde}
}
\keyword{datasets}
