% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\name{detect_low_expression}
\alias{detect_low_expression}
\title{Filter isoforms by absolute expression level across cell types}
\usage{
detect_low_expression(
  data,
  id_table,
  mean_counts = 10,
  median_counts = NULL,
  expressed_only = TRUE,
  isoform_col = NULL
)
}
\arguments{
\item{data}{A data.frame or tibble object including isoforms as rows and cells as columns.
Isoform IDs can be included as row names (data.frame) or as an additional column (tibble).}

\item{id_table}{A data frame including two columns named \code{cell} and \code{cell_type},
in which correspondence between cell ID and cell type should be
provided. The number of rows should be equal to the total number of
cell columns in \code{data}, and the order of the \code{cell} column should
match column (i.e. cell) order in \code{data}.}

\item{mean_counts}{A numeric value indicating the mean expression
threshold used to flag an isoform as lowly expressed. Defaults to 10 counts,
i.e. all isoforms with mean expression below 10 counts in all cell types will
be reported as lowly expressed.}

\item{median_counts}{A numeric value supplying a median expression threshold.
If provided, will override \code{mean_counts}, and the median expression of
the transcripts across cell types will be used to flag low expression.}

\item{expressed_only}{Logical. When \code{TRUE}, zero-expression cells will
not be considered to compute mean/median cell type expression. Transcripts
will therefore be flagged based solely on their count no. across cells
where they are expressed.}

\item{isoform_col}{When a tibble is provided in \code{data}, a character object
indicating the name of the column in which isoform IDs are specified.
Otherwise, isoform identifiers will be assumed to be defined as rownames,
and this argument will not need to be provided.}
}
\value{
A \code{tibble} containing two columns, the first one including
transcript IDs, and the second containing logical values specifying whether
the isoform was flagged as lowly expressed (considering the provided mean
or median expression threshold).
}
\description{
Filter isoforms by absolute expression level across cell types
}
