% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\name{detect_sparse}
\alias{detect_sparse}
\title{Filter isoforms by maximum cell type-level proportion of zeros}
\usage{
detect_sparse(data, id_table, ct_proportion = 0.2, isoform_col = NULL)
}
\arguments{
\item{data}{A data.frame or tibble object including isoforms as rows and cells as columns.
Isoform IDs can be included as row names (data.frame) or as an additional column (tibble).}

\item{id_table}{A data frame including two columns named \code{cell} and \code{cell_type},
in which correspondence between cell ID and cell type should be
provided. The number of rows should be equal to the total number of
cell columns in \code{data}, and the order of the \code{cell} column should
match column (i.e. cell) order in \code{data}.}

\item{ct_proportion}{A numeric indicating the minimum proportion of cells with
non-zero expression that will be allowed per cell type. Isoforms with a non-zero
value proportion below the threshold in all cell types will be flagged
as sparse. Defaults to 0.2 (i.e. 20\%).}

\item{isoform_col}{When a tibble is provided in \code{data}, a character object
indicating the name of the column in which isoform IDs are specified.
Otherwise, isoform identifiers will be assumed to be defined as rownames,
and this argument will not need to be provided.}
}
\value{
A \code{tibble} containing two columns, the first one including
transcript IDs, and the second containing logical values specifying whether
the isoform was flagged as sparse (considering the provided threshold of
cell type proportion with non-zero expression).
}
\description{
For convenience, \code{acorde} includes a function to assist
lenient filtering of isoforms based on the proportion of zero expression values
across cell types. Based on this criteria, a minimum number of cells must have
non-zero expression in at least one cell type.
}
