% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_refinement.R
\name{expand_clusters}
\alias{expand_clusters}
\title{Expand clusters with unclustered isoforms}
\usage{
expand_clusters(
  data,
  isoform_col = NULL,
  id_table,
  cluster_list,
  unclustered,
  percentile_no = 10,
  force_expand = TRUE,
  expand_threshold = NULL,
  method = c("percentile", "pearson", "spearman", "rho", "zi_kendall")
)
}
\arguments{
\item{data}{A data.frame or tibble object including isoforms as rows and
cells as columns. Isoform IDs can be included as row names (data.frame)
or as an additional column (tibble).}

\item{isoform_col}{When a tibble is provided in \code{data}, a character value
indicating the name of the column in which isoform IDs are specified.}

\item{id_table}{A data frame including two columns named \code{cell} and
\code{cell_type}, in which correspondence between cell ID and cell type should be
provided. The number of rows should be equal to the total number of
cell columns in \code{data}, and the order of the \code{cell} column should
match column (i.e. cell) order in \code{data}.}

\item{cluster_list}{A list of character vectors, each containing the
identifiers of the isoforms in a cluster.}

\item{unclustered}{A character vector containing the identifiers of
unclustered isoforms.}

\item{percentile_no}{Integer indicating the number of percentiles that will
be used to summarized cell type expression via \code{\link{percentile_expr}}.
Should always be higher than 4 (quantiles) and lower than 100 (percentiles).
Defaults to 10.}

\item{force_expand}{Logical. When \code{TRUE}, forced expansion is enabled,
and unclustered isoforms are assigned to clusters based on the maximum
reported co-expression. If \code{FALSE}, a co-expression threshold needs to
be specified via \code{expand_threshold}.}

\item{expand_threshold}{A numeric value defining the minimum correlation
required to assign an unclustered isoform to a cluster.}

\item{method}{Character indicating a co-expression method to use for merging
similar clusters. Should be one of \code{percentile, pearson, spearman,
zi_kendall, rho} (see details).}
}
\value{
A list containing expanded clusters, where each element will include
the identifiers of the isoforms assigned ot each cluster.
If \code{force_expand = FALSE}, the first element of the list will contain
the identifiers of isoforms that remained unassigned.
}
\description{
Assign unclustered isoforms based on their correlation with
the average expression profile of previously-generated clusters.
}
\details{
Correlation-based cluster expansion first requires cluster
\emph{metatranscripts} to be calculated. A cluster's metatranscript
is calculated as the mean of the
\link[acorde:percentile_expr]{percentile-summarized}
expression of all of the isoforms in that cluster. Next, co-expression values
between the metatranscripts and the unclustered isoforms are computed using
the similarity metric specified in \code{method}.

Available \strong{co-expression metrics} (selected via the \code{method}) include:
\enumerate{
   \item \code{percentile}: percentile correlations computed using
   \code{\link{percentile_cor}}.

   \item \code{pearson}: Pearson correlation computed using
   \code{\link[stats]{cor}}.

   \item \code{spearman}: Spearman correlation computed using
   \code{\link[stats]{cor}}.

   \item \code{zi_kendall}: zero-inflated Kendall correlation computed
   using the \code{\link[dismay]{dismay}} function.

   \item \code{rho}: rho proportionality metric computed using the
   \code{\link[dismay]{dismay}} function.
}

Cluster expansion can be forced or threshold-based, depending on the values
provided to \code{force_expansion} (i.e. \code{TRUE} or \code{FALSE},
respectively). If forced, isoforms will be assigned to a cluster if its
metatranscript is the one yielding the highest co-expression with the isoform's
expression, regardless of the exact co-expression metric value. Conversely,
when threshold-based, only isoforms showing co-expression above the
user-defined threshold (\code{expand_threshold}) with at least one
metatranscript will be assigned to clusters. In this case, the maximally-
correlated cluster is selected as the best match if there are several
candidate clusters with above-threshold co-expression values.
}
\references{
\insertRef{Skinnider2019}{acorde}

\insertRef{Venables2002}{acorde}
}
