% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_refinement.R
\name{filter_clusters}
\alias{filter_clusters}
\alias{single_cluster_filter}
\title{Remove isoforms with low correlation from clusters}
\usage{
filter_clusters(
  cluster_list,
  contains_unclustered = TRUE,
  cor_matrix,
  min_cor = 0.8,
  lowcor_threshold = 3,
  size_filter = TRUE,
  size_threshold = 3
)

single_cluster_filter(cluster, cor_matrix, min_cor, lowcor_threshold)
}
\arguments{
\item{cluster_list}{A list of character vectors, each containing the
identifiers of the isoforms in a cluster.}

\item{contains_unclustered}{A logical, indicating whether the first element
in \code{cluster_list} corresponds to unassigned isoforms
(i.e. unclustered, see \code{\link{cluster_isoforms}}).}

\item{cor_matrix}{A matrix including co-expression values (e.g.
correlations) between isoforms, where column and row names indicate isoform
IDs.}

\item{min_cor}{A numeric value indicating the minimum correlation
(or co-expression value) allowed between isoforms of the same cluster.}

\item{lowcor_threshold}{An integer indicating the number of correlation values
below \code{min_cor} allowed per isoform in the cluster. As a result, any
isoform with \code{correlation < min_cor} with \code{>= lowcor_threshold}
counterparts will be moved to the unclustered group.}

\item{size_filter}{A logical indicating whether size filter of the correlation-filtered
filtered clusters should be performed.
Size filter removes clusters that contain too few elements after filtering
and moves their isoforms to the unclustered group.}

\item{size_threshold}{An integer value indicating the minimum number of
isoforms required to preserve the cluster during size filtering.}
}
\value{
A list of clusters, where the first element will correspond to
unclustered isoforms and the rest to the correlation-filtered clusters.
}
\description{
Refine the expression signal of a set of isoform clusters
by removing elements with low correlations with the rest of isoforms in the
cluster.
}
\section{Functions}{
\itemize{
\item \code{single_cluster_filter}: applies the filtering parameters defined in
\code{filter_clusters} to one cluster. This function is called internally and
used to iterate and filter all clusters supplied via the \code{cluster_list}
argument.
}}

