% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fromList.R
\name{fromList}
\alias{fromList}
\title{Obtaining intersections from a list of named vectors}
\usage{
fromList(list)
}
\arguments{
\item{list}{A list of named vectors, each including one group of elements of
interest.}
}
\value{
A data.frame with the (unique) elements within the list as row names
and one column per group. Values of 0 or 1 will be shown if the element is
present or absent from the group, respectively.
}
\description{
Modified version of \code{\link[UpSetR]{fromList}}
(\code{UpSetR} package). Transforms a list of named vectors to a data frame
compatible with the \code{\link[UpSetR]{upset}} function.
}
\references{
The \code{fromList()} function is originally included in the
\code{\link[UpSetR:fromList]{UpSetR}} package:

\insertRef{Conway2017}{acorde}

The present modifications to the fromList() function were obtained from
user @docmanny's contribution in this
\href{https://github.com/hms-dbmi/UpSetR/issues/85#issuecomment-327900647}{GitHub issue}.
}
