% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gene_tr_ID}
\alias{gene_tr_ID}
\title{Gene-isoform associations for the mouse neural long read transcriptome}
\format{
A tibble containing the columns \code{transcript_id} and \code{gene_id}.
Isoform (i.e. transcript) identifiers correspond to those obtained during
long read-transcriptome definition (see Arzalluz Luque et al.)
}
\usage{
gene_tr_ID
}
\description{
Table associating long read-defined isoform IDs to their genes.
}
\references{
Long read transcriptome definition details can be found in the methods section of
Arzalluz-Luque et al.:

\insertRef{Arzalluz-Luque2021}{acorde}
}
\keyword{datasets}
