% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_genes.R
\name{make_design}
\alias{make_design}
\title{Create design matrix for co-DIU test of a pair of genes}
\usage{
make_design(pair, data, cluster_list, gene_tr_table, cell_types)
}
\arguments{
\item{pair}{A character vector of length 2 indicating the gene IDs of
a gene pair to test.}

\item{data}{A data.frame or tibble including isoforms as rows and cells
as columns. Isoform IDs can be included as row names (data.frame) or as an
additional column (tibble).}

\item{cluster_list}{A list of character vectors containing isoform IDs.
Each element of the list represents a cluster of isoforms.}

\item{gene_tr_table}{A data.frame or tibble containing two columns
named \code{transcript_id} and \code{gene_id}, indicating gene-isoform
correspondence.}

\item{cell_types}{A character vector including cell type assignments for
the cell IDs in \code{data}.}
}
\value{
A tibble, containing a long-form table with the required factors for
GLM fitting and statistical testing.
}
\description{
This function is used by \code{\link{test_shared_genes}} to
create a design matrix for GLM fitting using information from a pair of co-DIU
candidategenes, i.e. expression values of the gene's co-expressed isoforms,
cell type labels for cells and cluster labels for isoforms.
}
\details{
Arguments are passed from \code{\link{test_shared_genes}}.
}
