% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_genes.R
\name{make_test}
\alias{make_test}
\title{Perform co-DIU statistical test for one gene pair}
\usage{
make_test(pair, data, cluster_list, gene_tr_table, cell_types)
}
\arguments{
\item{pair}{A character vector of length 2 indicating the gene IDs of
a gene pair to test.}

\item{data}{A data.frame or tibble including isoforms as rows and cells
as columns. Isoform IDs can be included as row names (data.frame) or as an
additional column (tibble).}

\item{cluster_list}{A list of character vectors containing isoform IDs.
Each element of the list represents a cluster of isoforms.}

\item{gene_tr_table}{A data.frame or tibble containing two columns
named \code{transcript_id} and \code{gene_id}, indicating gene-isoform
correspondence.}

\item{cell_types}{A character vector including cell type assignments for
the cell IDs in \code{data}.}
}
\value{
A tibble containing two columns, \code{cluster:cell_type} and
\code{gene:cell_type}, containing the p-value obtained in the likelihood ratio
chi-square test (see \code{\link{test_shared_genes}}) for each of these two
interactions.
}
\description{
This function is called iteratively by \code{\link{test_shared_genes}}
to test cluster-dependent expression for a single pair of candidate co-DIU genes.
}
\details{
Arguments are passed from \code{\link{test_shared_genes}}.
}
\references{
\insertRef{Venables2002}{acorde}

\insertRef{Fox2019}{acorde}
}
