% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{merge_clusters}
\alias{merge_clusters}
\title{Merge redundant clusters by expression profile similarity}
\usage{
merge_clusters(
  data,
  isoform_col = NULL,
  id_table,
  cluster_list,
  percentile_no = 10,
  dynamic = FALSE,
  method = c("percentile", "pearson", "spearman", "rho", "zi_kendall"),
  height_cutoff = 0.2,
  cutree_no = NULL,
  ...
)
}
\arguments{
\item{data}{A data.frame or tibble object including isoforms as rows and
cells as columns. Isoform IDs can be included as row names (data.frame)
or as an additional column (tibble).}

\item{isoform_col}{When a tibble is provided in \code{data}, a character value
indicating the name of the column in which isoform IDs are specified.}

\item{id_table}{A data frame including two columns named \code{cell} and
\code{cell_type}, in which correspondence between cell ID and cell type should be
provided. The number of rows should be equal to the total number of
cell columns in \code{data}, and the order of the \code{cell} column should
match column (i.e. cell) order in \code{data}.}

\item{cluster_list}{A list of character vectors, each containing the
identifiers of the isoforms in a cluster.}

\item{percentile_no}{Integer indicating the number of percentiles that will
be used to summarized cell type expression via \code{\link{percentile_expr}}.
Should always be higher than 4 (quantiles) and lower than 100 (percentiles).
Defaults to 10.}

\item{dynamic}{A logical. If \code{TRUE}, merge will be performed
via dynamic hierarchical clustering. Defaults to \code{FALSE}.}

\item{method}{Character indicating a co-expression method to use for merging
similar clusters. Should be one of \code{percentile, pearson, spearman,
zi_kendall, rho} (see details). Percentile correlation is used by default.}

\item{height_cutoff}{When \code{dynamic = FALSE}, a numeric value between
0 and 1 to be supplied to \code{\link[stats]{cutree}} via the \code{h} argument.
Indicates the height where the created dendrogram tree should be cut to
generate groups of merged clusters.}

\item{cutree_no}{An integer indicating the desired number of groups to
merge clusters into. Supplied to \code{\link[stats]{cutree}} via the
\code{k} argument. Only required when \code{dynamic = FALSE} and
\code{height_cutoff = NULL}.}

\item{...}{Additional arguments passed to \code{\link[dynamicTreeCut]{cutreeHybrid}}
(only when \code{dynamic = TRUE}).}
}
\value{
A named list containing two elements:

\enumerate{
   \item \code{merged_groups}: a list detailing merge decisions, in which
   each element contains the identifiers of the clusters that were merged together.
   \item \code{clusters}: a list of character vectors, containing the
   identifiers of isoforms included in each of the resulting clusters.
}
}
\description{
Join clusters representing the same expression
pattern across cell types (redundant clusters). This function uses a
metaclustering system (see details) and user-defined similarity thresholds
that allows to control for the stringency of the merge process.
}
\details{
During the isoform clustering process, it is generally useful to
prioritize the reduction of within-cluster variability. This, however, can lead
to obtaining a large number of small, redundant clusters. To mitigate this
effect, \code{acorde} includes a step where clusters with high profile similarity
can be merged using the correlation between their \emph{metatranscripts}.
A cluster's metatranscript is calculated as the mean of the
\link[acorde:percentile_expr]{percentile-summarized}
expression of all of the isoforms in that cluster. Then, co-expression values
between metatranscripts are calculated and used to
generate a distance matrix to group cluster profiles by similarity, a process
that can be referred to as \emph{metaclustering}.

By default, the \strong{metaclustering} proccess is done using traditional
hierarchical clustering via \code{\link[stats]{hclust}},
which requires the definition of either a height cutoff (\code{height_cutoff}
parameter) or a number of clusters to obtain (\code{cutree_no}).

Available \strong{co-expression metrics} (selected via the \code{method}) include:
\enumerate{
   \item \code{percentile}: percentile correlations computed using
   \code{\link{percentile_cor}}.

   \item \code{pearson}: Pearson correlation computed using
   \code{\link[stats]{cor}}.

   \item \code{spearman}: Spearman correlation computed using
   \code{\link[stats]{cor}}.

   \item \code{zi_kendall}: zero-inflated Kendall correlation computed
   using the \code{\link[dismay]{dismay}} function.

   \item \code{rho}: rho proportionality metric computed using the
   \code{\link[dismay]{dismay}} function.
}

Alternatively, users may choose to perform metatranscript clustering
dynamically using the \code{dynamicTreeCut} package, therefore setting \code{dynamic = TRUE}.
In this case, additional parameters will need to be supplied to the
\code{\link[dynamicTreeCut]{cutreeHybrid}} function via the \code{...} argument.
Note that \code{minClusterSize = 1} is set internally to allow clusters to
remain unmerged if no redundancies with the profiles of other clusters are
found.
}
\references{
\insertRef{Langfelder2008}{acorde}

\insertRef{Venables2002}{acorde}

\insertRef{Skinnider2019}{acorde}
}
