% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metadata_down}
\alias{metadata_down}
\title{Tasic metadata (post-downsampling)}
\format{
A tibble. Contains 10 columns, including cell IDs (taken from
sequencing run IDs in the \code{run} column), cell types and subtypes assgined
by the authors (\code{cell_type} and \code{subtype} columns).
}
\usage{
metadata_down
}
\description{
Metadata table provided from publicly available experimental
details from Tasic et al. (see references).
}
\references{
Public single-cell RNA-Seq data (Smart-Seq2) from mouse visual cortex was
obtained from Tasic et al.:

\insertRef{Tasic2016a}{acorde}
}
\keyword{datasets}
