% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percentile_cor.R
\name{percentile_cor}
\alias{percentile_cor}
\title{Compute percentile correlations between a set of transcripts}
\usage{
percentile_cor(data, id_table, percentile_no = 10, isoform_col = NULL)
}
\arguments{
\item{data}{A data.frame or tibble object including isoforms as rows and
cells as columns. Isoform IDs can be included as row names (data.frame)
or as an additional column (tibble).}

\item{id_table}{A data frame including two columns named \code{cell} and
\code{cell_type}, in which correspondence between cell ID and cell type should be
provided. The number of rows should be equal to the total number of
cell columns in \code{data}, and the order of the \code{cell} column should
match column (i.e. cell) order in \code{data}.}

\item{percentile_no}{Integer indicating the number of percentiles to generate
for each of the cell types. Should always be higher than 4 (quantiles)
and lower than 100 (percentiles). Defaults to 10.}

\item{isoform_col}{When a tibble is provided in \code{data}, a character value
indicating the name of the column in which isoform IDs are specified.}
}
\value{
A correlation matrix including one column-row pair per isoform
included in \code{data}.
}
\description{
This function summarizes expression for each cell type
using percentiles and then calculates Pearson correlations between all possible
transcript pairs in the data set. Internally, the function function calls
\code{\link{percentile_expr}} to generate percentile-summarized expression, and
then \code{\link[stats]{cor}} to compute Pearson correlation
}
