% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percentile_cor.R
\name{percentile_expr}
\alias{percentile_expr}
\title{Summarize cell type-level expression using percentiles}
\usage{
percentile_expr(data, id_table, percentile_no = 10, isoform_col = NULL)
}
\arguments{
\item{data}{A data.frame or tibble object including isoforms as rows and
cells as columns. Isoform IDs can be included as row names (data.frame)
or as an additional column (tibble).}

\item{id_table}{A data frame including two columns named \code{cell} and
\code{cell_type}, in which correspondence between cell ID and cell type should be
provided. The number of rows should be equal to the total number of
cell columns in \code{data}, and the order of the \code{cell} column should
match column (i.e. cell) order in \code{data}.}

\item{percentile_no}{Integer indicating the number of percentiles to generate
for each of the cell types. Should always be higher than 4 (quantiles)
and lower than 100 (percentiles). Defaults to 10.}

\item{isoform_col}{When a tibble is provided in \code{data}, a character value
indicating the name of the column in which isoform IDs are specified.}
}
\value{
A \code{\link[tibble]{tibble}} containing one column of percentile-
summarized expression values per input transcript
in \code{data}.
}
\description{
For each of the isoforms in the single-cell data set,
the function calculates percentile values using cell type-level expression values,
providing an estimate of the expression distribution shown by a particular
isoform in each of the cell types.
}
