% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_visualization.R
\name{plot_cluster_ctmeans}
\alias{plot_cluster_ctmeans}
\title{Plot cell type expression of all isoforms in cluster}
\usage{
plot_cluster_ctmeans(data, plot_title = NULL, ct_labels = NULL)
}
\arguments{
\item{data}{A tibble including cell type-level mean expression for all isoforms
in a cluster (output by \code{\link{calculate_cluster_ctmeans}}).}

\item{plot_title}{A character object providing a title for the plot.}

\item{ct_labels}{A character vector including a plot label for each of
the cell types defined in \code{data}.}
}
\value{
A plot object generated by \code{\link[ggplot2]{ggplot}}.
}
\description{
This function uses the output of \code{\link{calculate_cluster_ctmeans}}
to create a plot including all of the isoforms assigned to a cluster, enabling comparison
of their expression pattern across cell types.
}
