% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_visualization.R
\name{plot_cluster_profile}
\alias{plot_cluster_profile}
\title{Plot mean profile of an isoform cluster}
\usage{
plot_cluster_profile(data, plot_title = NULL, ct_labels = NULL)
}
\arguments{
\item{data}{A tibble including a global mean expression value for each cell type,
computed across all isoforms in the cluster (output by
\code{\link{calculate_cluster_ctmeans}}).}

\item{plot_title}{A character object providing a title for the plot.}

\item{ct_labels}{A character vector including a plot label for each of
the cell types defined in \code{data}.}
}
\value{
A plot object generated by \code{\link[ggplot2]{ggplot}}.
}
\description{
This function uses the output of \code{\link{calculate_cluster_profile}}
to plot the mean scaled expression of all isoforms in a cluster, enabling a
summarized vision of the cluster's global pattern.
}
