% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_type_DE.R
\name{run_downsampling}
\alias{run_downsampling}
\title{Run one downsampling iteration on Tasic et al. dataset}
\usage{
run_downsampling(data, id_table, downsampling_ct, cell_no, isoform_col = NULL)
}
\arguments{
\item{data}{A data.frame or tibble object including isoforms as rows and cells
as columns. Isoform IDs should be included in an independent column, not defined
as \code{rownames}.}

\item{id_table}{A data frame including two columns named \code{cell} and
\code{cell_type}, in which correspondence between cell ID and cell type should be
provided.}

\item{downsampling_ct}{A character vector including one or more cell type names
(matching those in \code{id_table$cell_type}) to be targeted by downsampling.}

\item{cell_no}{A numeric indicating the number of cells to be randomly sampled
during downsampling. Should be the same for all targeted cell types.}

\item{isoform_col}{Name of the column in \code{data} that contains isoform IDs.
Otherwise, isoform identifiers will be assumed to be defined as rownames,
and this argument will not need to be provided.}
}
\value{
An SCE object containing the data after downsampling in
\code{assay(counts = data)} and \code{id_table} as
metadata in the \code{colData()} slot. This format corresponds to the input of
\code{\link{cell_type_DE}}.
}
\description{
Function designed to pack random cell downsampling code to ease
running multiple iterations of this process during \code{acorde} benchmarking.
Each of these iterations includes three steps: randomly selecting a number of
cells form the specified cell types, subsetting the expression matrix and
creating an SCE object including ZINBWaVE weights for DE testing
(see \code{\link{cell_type_DE}}).
}
