% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coexpr_simulation.R
\name{shuffle_group_matrix}
\alias{shuffle_group_matrix}
\title{Reorder cell type-specific expression matrix during co-expression simulation}
\usage{
shuffle_group_matrix(sim_data, feature_ids, group_pattern, cluster_no)
}
\arguments{
\item{sim_data}{A count matrix with features as rows and cells as columns.
Feature IDs must be included in an additional column named \code{feature}.}

\item{feature_ids}{A two-column \code{tibble} containing \code{top} and \code{bottom}
columns, each including the feature IDs of features to be used as highly or
lowly expressed when shuffling by the indicated expression pattern.}

\item{group_pattern}{A logical vector, containing \code{TRUE} to indicate that
high expression in that cell type is desired and \code{FALSE} if the opposite.
The vector must be ordered as the cell types in \code{sim_data}.}

\item{cluster_no}{An integer indicating the total number of co-expression
clusters that are being generated in the simulation}
}
\value{
An expression matrix, with the same characteristics as \code{sim_data},
and a number of features defined as the total amount of top/bottom features
selected divided by the number of clusters for which co-expression patterns
where supplied.
}
\description{
This function is used internally by \code{acorde} to perform
the shuffling of simulated features for an individual cell type, as part of
the co-expression simulation process. The function is called recursively by
\code{\link[acorde:simulate_coexpression]{simulate_coexpression()}} to
perform the simulation on a full scRNA-seq matrix.
}
