% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coDIU_genes.R
\name{test_shared_genes}
\alias{test_shared_genes}
\title{Statistical testing of candidate co-DIU genes}
\usage{
test_shared_genes(
  data,
  cluster_list,
  shared_genes,
  gene_tr_table,
  id_table,
  isoform_col = NULL
)
}
\arguments{
\item{data}{A data.frame or tibble object including isoforms as rows and cells
as columns. Isoform IDs can be included as row names (data.frame) or as an
additional column (tibble).}

\item{cluster_list}{A list of character vectors containing isoform IDs.
Each element of the list represents a cluster of isoforms.}

\item{shared_genes}{A two-row matrix containing \emph{n} candidate co-DIU
gene pairs as column. Typically the result of running
\code{\link{find_shared_genes}}.}

\item{gene_tr_table}{A data.frame or tibble object containing two columns
named \code{transcript_id} and \code{gene_id}, indicating gene-isoform
correspondence.}

\item{id_table}{A data frame including two columns named \code{cell}
and \code{cell_type}, in which correspondence between cell ID and cell type
should be provided. The number of rows should be equal to the total number of
cell columns in \code{data}, and the order of the \code{cell} column should
match column (i.e. cell) order in \code{data}.}

\item{isoform_col}{When a tibble is provided in \code{data}, a character
object indicating the name of the column where isoform IDs are specified.
Otherwise, isoform identifiers will be assumed to be defined as rownames,
and this argument will not need to be provided.}
}
\value{
A list containing one \code{tibble} per tested gene pair, as generated
by \code{\link{make_test}}. Each tibble will include two columns,
\code{cluster:cell_type} and \code{gene:cell_type}, containing the p-value
obtained when testing each of these interactions in the type-II
ANOVA test.

\strong{NOTE:} In some cases the assumptions required for fitting the GLM are not met,
and an \code{NA} value is returned instead. These are output to allow users
to control for untested gene pairs, but can easily be removed from the output.
}
\description{
Pairwise statistical testing of co-Differential Isoform Usage
relationships. For a selected set of gene pairs showing co-expression of
isoforms across clusters (see \code{\link{find_shared_genes}}), this
function tests the significance of the detected co-DIU patterns.
}
\details{
A set of \strong{potentially co-DIU genes} will have at least two of their
isoforms assigned to the same clusters, i.e. show detectable
isoform-level co-expression. However, since clustering allows isoforms
with slightly variable expression patterns to be clustered together,
some isoforms might be assigned to clusters that do not faithfully represent their
expression profile, leading to inaccuracies in co-DIU detection.
To \strong{avoid false-positive co-DIU genes}, the present function
applies a regression model and a statistical
test to each of the candidate pair of genes (hereby named
gene 1 and gene 2), where at least two of the isoforms of each gene must
belong to the same two clusters (hereby named cluster 1 and cluster 2).

Briefly, we need to assess whether expression values for the isoforms follow
a correct co-DIU pattern, that is, the average profile across cell types of
the two isoforms in cluster 1 must be significantly different to the average
profile of the two isoforms in cluster 2, indicating distinct expression
profiles for the two isoforms of each gene. In addition, the average profile
of the two isoforms of gene 1 must not be different to the average profile of
the two isoforms of gene 2, indicating that co-expression is only detectable
when isoform-level expression is considered.

Internally, the function fits a \strong{generalized linear regression model} (GLM) via
the \code{\link[stats]{glm}} function, using the \code{\link[MASS]{negative.binomial}}
function in the \code{MASS} package to set the error distribution and link
function of the model via the \code{family} argument. To test the
significance of the \emph{cluster*cell type}
and \emph{gene*cell type} interactions (as described above), we calculated
\strong{type-II analysis-of-variance} (ANOVA) tables for the model using a
\strong{likelihood-ratio chi-square test} using the \code{\link[car]{Anova}} function in
the \code{car} package (given the unbalanced design).
}
\references{
\insertRef{Venables2002}{acorde}

\insertRef{Fox2019}{acorde}
}
\seealso{
For details, see internal functions:
\code{\link{make_design}}, \code{\link{make_test}}.
}
