% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_type_DE.R
\name{cell_type_DE}
\alias{cell_type_DE}
\title{Test differential expression of isoforms across cell types}
\usage{
cell_type_DE(
  data,
  mode = c("edgeR", "DESeq2", "both"),
  compute_weights = TRUE,
  AdjPvalue = NULL,
  offset = NULL
)
}
\arguments{
\item{data}{A \code{\link[SingleCellExperiment]{SingleCellExperiment}} (SCE)
object including isoform-level counts on the \code{assay()} slot and a
\code{cell_type} column indicating cell type correspondence in the
\code{colData()} slot.}

\item{mode}{A character value indicating whether \code{"edgeR"},
\code{"DESeq2"} or \code{"both"} DE methods are to be run.}

\item{compute_weights}{A logical value. If \code{TRUE}, cell-level weights
will be computed using ZINBWaVE (see \code{\link[zinbwave]{zinbwave}}).
Alternatively, weights may be pre-computed and stored in the \code{weights()}
slot in the SCE object.}

\item{AdjPvalue}{A numeric value. If provided, filtering of DE results by
adjusted p-value is enabled.}

\item{offset}{Optionally, a matrix containing offset information to be
considered for bias correction in edgeR and/or DESeq2 DE tests.}
}
\value{
If \code{mode = "both"}, a named list containing the results (filtered
or not, depending on whether \code{AdjPvalue} is provided) of testing
Differential Expression using edgeR and DESeq2. If the \code{mode} argument
was used to select only one method, the function returns a data.frame object
containing the results table obtained by that method.
}
\description{
This function provides a wrapper to compute Differential
Expression across multiple cell types on a single-cell dataset, using
standard bulk tools edgeR and DESeq2 and ZINBWaVE weights, as described
by Van den Berge et al. (see references below).
}
\references{
\insertRef{Robinson2009}{acorde}

\insertRef{Love2014}{acorde}

\insertRef{VandenBerge2018}{acorde}
}
