% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_type_DE.R
\name{detect_sparse}
\alias{detect_sparse}
\title{Filter isoforms by maximum cell type-level proportion of zeros}
\usage{
detect_sparse(data, cell_types, ct_proportion = 0.2, isoform_col = NULL)
}
\arguments{
\item{data}{A data.frame or tibble object including isoforms as rows and cells as columns.
Isoform IDs can be included as row names (data.frame) or as an additional column (tibble).}

\item{cell_types}{A character vector including cell type information for all the
cells in \code{data}, in the right order. Vector length should be equal to
the total number of cell columns in \code{data}.}

\item{ct_proportion}{A numeric indicating the minimum proportion of cells with
non-zero expression that will be allowed per cell type. Isoforms with a non-zero
value proportion above the threshold in at least one cell type will be flagged
to be preserved. Defaults to 0.2 (i.e. 20%).}

\item{rownames}{If a tibble is provided in \code{data} a character value
indicating the name of the column in which isoform IDs are specified.}
}
\value{
A logical vector including one entry per isoform in \code{data}. Isoforms
meeting the sparsity criteria will have a value of \code{TRUE}, and otherwise
be labeled as \code{FALSE}. This logical vector can then be used to filter
isoforms, i.e. the rows in \code{data}.
}
\description{
For convenience, \code{acorde} includes a function to assist
lenient filtering of isoforms based on the proportion of zero expression values
across cell types. Based on this criteria, a minimum number of cells must have
non-zero expression in at least one cell type.
}
