% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_refinement.R
\name{keep_DIU}
\alias{keep_DIU}
\title{Keep isoforms from genes with Differential Isoform Usage}
\usage{
keep_DIU(cluster_list, gene_tr_table)
}
\description{
Keep isoforms from genes with Differential Isoform Usage
}
