% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_genes.R
\name{test_shared_genes}
\alias{test_shared_genes}
\title{Statistical testing of candidate co-DIU genes}
\usage{
test_shared_genes(data, cluster_list, shared_genes, gene_tr_table, cell_types)
}
\description{
Pairwise statistical testing of co-Differential Isoform Usage
relationships. For a selected set of gene pairs showing co-expression of
isoforms across clusters (see \code{\link{find_shared_genes}}), this
function tests the significance of the detected co-DIU patterns.
}
\details{
A set of potentially co-DIU genes will have at least two of their
isoforms assigned to the same clusters, i.e. show detectable
isoform-level co-expression. However, since clustering allows isoforms
with slightly variable expression patterns to be clustered together,
some isoforms might be assigned to clusters that do not faithfully represent their
expression profile, leading to detection of false-positive co-DIU genes.
To avoid this, the present function applies a regression model and a statistical
test to each of the candidate pair of genes (hereby named
gene 1 and gene 2), where at least two of the isoforms of each gene must
belong to the same two clusters (hereby named cluster 1 and cluster 2).

Briefly, we need to assess whether expression values for the isoforms follow
a correct co-DIU pattern, that is, the average profile across cell types of
the two isoforms in cluster 1 must be significantly different to the average
profile of the two isoforms in cluster 2, indicating distinct expression
profiles for the two isoforms of each gene. In addition, the average profile
of the two isoforms of gene 1 must not be different to the average profile of
the two isoforms of gene 2, indicating that co-expression is only detectable
when isoform-level expression is considered.

Internally, the function fits a generalized linear regression model (GLM) using
the \code{\link[stats]{glm}} function, using the \code{\link[MASS]{negative.binomial}}
function in the \code{MASS} package to set the error distribution and link
function of the model via the \code{family} argument. To test the
significance of the \emph{cluster*cell type}
and \emph{gene*cell type} interactions (as described above), we calculated
type-II analysis-of-variance (ANOVA) tables for the model using a
likelihood-ratio chi-square test using the \code{\link[car]{Anova}} function in
the \code{car} package (given the unbalanced design).
}
\references{
\insertRef{Venables2002}{acorde}

\insertRef{Fox2019}{acorde}
}
